cansat/backend/publisher/requirements.txt
pyserial==3.5
redis==4.5.4
python-dotenv==1.0.0
requests==2.32.5